refactor: restructure graphics items and add RectangleItem class
